+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Renamed all `Retry.when` methods to `Retry.until` to better emphasize what the parameters actually do.
 - Renamed `Accumulator.emptyAccumulator` method to `Accumulator.empty` for brevity.
+- Changed `Retry.until` methods to always have the first 2 parameters the result supplier first and the exit condition second where applicable.
 
 ---